requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
pandas>=2.0.0
pyyaml>=6.0
trino>=0.333.0
